introduce s8 type and use it to simplify theme parsing
Previously theme parsing included strlen(), even though we could
have retrieved the length directly from lua, and strdup()/free()
because we needed to split the string which required inserting a 0
for termination. This was all energy wasting busy work since all
the information we actually needed was contained in the original
string.

This commit adds ~100 lines of code because I have provided my own
implementation of some functions that we were previously provided
by the C-runtime. This code will be reused later to simplify lots
of other code in sam.c, map.c, vis-lua.c, and likely more. I can't
say if this will be a net reduction of code in vis but I can
guarantee that it will be a net reduction in total lines of code
running on the users machine - C-runtime implementations are often
massive over-complicated messes (even musl).